+++ packages/compilers-types/src/SolidityTypes.ts
@@ -98,7 +98,7 @@ export interface SolidityJsonInput {
   settings: SoliditySettings;
 }
 
-interface SolidityOutputError {
+export interface SolidityOutputError {
   sourceLocation?: {
     file: string;
     start: number;
@@ -109,7 +109,9 @@ interface SolidityOutputError {
   severity: "error" | "warning";
   message: string;
   formattedMessage?: string;
+  errorCode?: string;
 }
+
 interface SolidityOutputEvmBytecode {
   object: string;
   opcodes?: string;
@@ -139,14 +141,14 @@ export interface SolidityOutputSources {
 }
 
 export interface StorageLayout {
-  storage: {
+  storage: Array<{
     astId: number;
     contract: string;
     label: string;
     offset: number;
     slot: string;
     type: string;
-  };
+  }>;
   types: {
     [index: string]: {
       encoding: string;