+++ packages/lib-sourcify/src/Compilation/CompilationTypes.ts
@@ -45,7 +45,9 @@ export type CompilationErrorCode =
 export class CompilationError extends SourcifyLibError {
   declare code: CompilationErrorCode;
   constructor(
-    params: SourcifyLibErrorParameters & { code: CompilationErrorCode },
+    params: SourcifyLibErrorParameters & {
+      code: CompilationErrorCode;
+    },
   ) {
     super(params);
   }