+++ services/server/src/server/apiv2/errors.ts
@@ -15,6 +15,7 @@ import {
 } from "@ethereum-sourcify/lib-sourcify";
 import { TooManyRequests } from "../../common/errors/TooManyRequests";
 import { BadGatewayError } from "../../common/errors/BadGatewayError";
+import { JobErrorData } from "../services/utils/database-util";
 
 export type ErrorCode =
   | VerificationErrorCode
@@ -22,6 +23,7 @@ export type ErrorCode =
   | "route_not_found"
   | "unsupported_chain"
   | "invalid_parameter"
+  | "invalid_json"
   | "proxy_resolution_error"
   | "job_not_found"
   | "duplicate_verification_request"
@@ -43,6 +45,7 @@ export interface MatchingErrorResponse extends GenericErrorResponse {
   onchainCreationCode?: string;
   onchainRuntimeCode?: string;
   creationTransactionHash?: string;
+  errorData?: JobErrorData;
 }
 
 export class InternalError extends InternalServerError {
@@ -188,6 +191,19 @@ export class MalformedEtherscanResponseError extends BadRequestError {
   }
 }
 
+export class InvalidJsonError extends BadRequestError {
+  payload: GenericErrorResponse;
+
+  constructor(message: string) {
+    super(message);
+    this.payload = {
+      customCode: "invalid_json",
+      message,
+      errorId: uuidv4(),
+    };
+  }
+}
+
 // Maps OpenApiValidator errors to our custom error format
 export function errorHandler(
   err: any,
@@ -198,6 +214,11 @@ export function errorHandler(
   // Let errors pass that already match the v2 error format
   if (err.payload) {
     next(err);
+    return;
+  }
+
+  if (err instanceof SyntaxError) {
+    next(new InvalidJsonError(err.message));
     return;
   }