+++ services/server/test/integration/apiv2/jobs.spec.ts
@@ -87,6 +87,7 @@ describe("GET /v2/verify/:verificationId", function () {
         recompiledRuntimeCode,
         onchainCreationCode,
         onchainRuntimeCode,
+        errorData,
       };
     }